fix: delete alert

+++ src/components/collection/delete-collection.js
@@ -1,6 +1,17 @@
 import React from 'react';
 import { useAdminDeleteCollection } from 'medusa-react';
 import { useToast } from '../ui/use-toast';
+import {
+    AlertDialog,
+    AlertDialogAction,
+    AlertDialogCancel,
+    AlertDialogContent,
+    AlertDialogDescription,
+    AlertDialogFooter,
+    AlertDialogHeader,
+    AlertDialogTitle,
+    AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
 
 const CollectionDeleteButton = ({ id }) => {
     const {toast} = useToast();
@@ -22,17 +33,32 @@ const CollectionDeleteButton = ({ id }) => {
     };
 
     return (
-        <div className="m-[-10px] p-[-10px]">
-            <div className="mt-[25px] flex justify-around">
-                <button className="h-[50px] w-4/5 rounded bg-red-500"
-                    onClick={() => {
-                        handleDelete()
-                    }}
-                >
-                    Có
-                </button>
-            </div>
-        </div>
+        <AlertDialog>
+            <AlertDialogTrigger className="w-full mx-auto mt-[2px] h-[30px] rounded bg-red-400 py-[4px] text-center align-middle">
+                    Xóa
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+                <AlertDialogHeader>
+                    <AlertDialogTitle>
+                        Bạn có muốn xóa bộ sưu tập này?
+                    </AlertDialogTitle>
+                    <AlertDialogDescription>
+                        ⚠️Thao tác này không thể hoàn lại!
+                    </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter className="inline-block">
+                    <AlertDialogCancel className="w-[49%] bg-gray-300 inline">Không</AlertDialogCancel>
+                    <AlertDialogAction
+                        onClick={() => {
+                            handleDelete();
+                        }}
+                        className="w-[49%] inline"
+                    >
+                        Có
+                    </AlertDialogAction>
+                </AlertDialogFooter>
+            </AlertDialogContent>
+        </AlertDialog>
     );
 };
